fix(8.3.10): 2. use rationals instead of reals
This is one option (which you've taken for most of the book, presumably
to avoid the whole noncomputable thing).

Alternatively, this could be a place to insert a footnote about the
noncomputability of division (and many other functions) over reals.

If you want to open the can of worms and mention the noncomputability,
these were a few useful links I found that discuss it:
https://leanprover.github.io/logic_and_proof/the_real_numbers.html
https://www.ma.imperial.ac.uk/~buzzard/xena/formalising-mathematics-2022/Part_A/section02reals/reals.html
https://leanprover-community.github.io/archive/stream/113489-new-members/topic/function.20definition.20noncomputable.html